Create Unofficial_TARDIS.txt
I have had an unofficial E2 chip with multiple features in the works for a long time and thought that it's probably ready for a release now.

IMPORTANT: THIS WILL MOST LIKELY ONLY WORK ON DR. MATT'S SERVER, THE ABYSS.

IP: 93.190.140.147:27015

https://github.com/Cookiemorph/Unofficial-TARDIS-E2/releases

There is a readme somewhere in my github that tells you where to put it. 

Comprehensive list of features i can remember:

>Fuel system, including refuelling
>Chance of paradox while refuelling
>Landing Lights
>Emergency Program 1
>Companion System
>Secure mode; Hologram guard and full lockdown
>TARDIS Visual Interface Call
>Autospin
>Skin change upon full repair
>Exterior cloister bell (sometimes glitchy)
>Self Destruct

REMEMBER:To use this, spawn the chip INSIDE the TARDIS, anywhere on the interior prop. Then type !help to see the commands.

Have fun.